fix: dtkcompressor docs index and dtkcompressor examples restore
dtkcompressor docs index and dtkcompressor examples restore

Log: dtkcompressor docs index and dtkcompressor examples restore
