+++ examples/dtkcompressor/CMakeLists.txt
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2023 UnionTech Software Technology Co., Ltd.
+#
+# SPDX-License-Identifier: LGPL-3.0-or-later
+
+cmake_minimum_required(VERSION 3.13)
+
+add_subdirectory(dtkcompressor)
