+++ examples/dtkcompressor/dtkcompressor/main.cpp
@@ -1,0 +1,20 @@
+// SPDX-FileCopyrightText: 2023 UnionTech Software Technology Co., Ltd.
+//
+// SPDX-License-Identifier: LGPL-3.0-or-later
+
+#include "compressortest.h"
+
+#include <QApplication>
+#include <QCommandLineParser>
+#include <QCommandLineOption>
+#include <QDir>
+
+int main(int argc, char *argv[])
+{
+    QApplication app(argc, argv);
+
+    CompressorTest compressor;
+
+    compressor.show();
+    return app.exec();
+}
